+++ src/vaultsieve/importers/__init__.py
@@ -1,4 +1,25 @@
 from vaultsieve.importers.bitwarden import import_bitwarden
 from vaultsieve.importers.csv_generic import import_csv
+from vaultsieve.importers.dashlane import import_dashlane
+from vaultsieve.importers.dashlane_json import import_dashlane_json
+from vaultsieve.importers.keepass import import_keepass
+from vaultsieve.importers.keepass_xml import import_keepass_xml
+from vaultsieve.importers.keeper import import_keeper
+from vaultsieve.importers.keeper_json import import_keeper_json
+from vaultsieve.importers.lastpass import import_lastpass
+from vaultsieve.importers.onepassword import import_onepassword
+from vaultsieve.importers.roboform import import_roboform
 
-__all__ = ["import_bitwarden", "import_csv"]
+__all__ = [
+    "import_bitwarden",
+    "import_csv",
+    "import_dashlane",
+    "import_dashlane_json",
+    "import_keepass",
+    "import_keepass_xml",
+    "import_keeper",
+    "import_keeper_json",
+    "import_lastpass",
+    "import_onepassword",
+    "import_roboform",
+]